Refactor ListCommand and add ListAllTagCommand
