+++ src/test/java/seedu/address/logic/LogicManagerTest.java
@@ -20,7 +20,7 @@ import org.junit.jupiter.api.io.TempDir;
 import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.logic.commands.musician.AddCommand;
-import seedu.address.logic.commands.musician.ListCommand;
+import seedu.address.logic.commands.general.ListCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
