+++ src/test/java/seedu/address/logic/commands/musician/ListCommandTest.java
@@ -8,6 +8,7 @@ import static seedu.address.testutil.typicalentities.TypicalAddressBook.getTypic
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
+import seedu.address.logic.commands.general.ListCommand;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
 import seedu.address.model.UserPrefs;
